refactor(backend): Move storable impls to correct file
